consolidate merkle objects
