+++ proto/src/domain/mod.rs
@@ -3,8 +3,7 @@ use miden_objects::{StarkField, Word};
 pub mod accounts;
 pub mod blocks;
 pub mod digest;
-pub mod merkle_path;
-pub mod mmr_delta;
+pub mod merkle;
 pub mod notes;
 pub mod nullifiers;
 pub mod transactions;
