+++ proto/src/domain/nullifiers.rs
@@ -1,12 +1,8 @@
-use miden_objects::{
-    crypto::merkle::{LeafIndex, MerklePath, SmtLeaf, SmtProof},
-    Digest, Word,
-};
+use miden_objects::{crypto::merkle::SmtProof, Digest};
 
 use crate::{
-    domain::{convert, try_convert},
     errors::{MissingFieldHelper, ParseError},
-    generated::{responses::NullifierBlockInputRecord, smt},
+    generated::responses::NullifierBlockInputRecord,
 };
 
 // NULLIFIER INPUT RECORD
@@ -40,97 +36,6 @@ impl From<NullifierInputRecord> for NullifierBlockInputRecord {
         Self {
             nullifier: Some(value.nullifier.into()),
             opening: Some(value.proof.into()),
-        }
-    }
-}
-
-impl TryFrom<smt::SmtLeaf> for SmtLeaf {
-    type Error = ParseError;
-
-    fn try_from(value: smt::SmtLeaf) -> Result<Self, Self::Error> {
-        let leaf = value.leaf.ok_or(smt::SmtLeaf::missing_field(stringify!(leaf)))?;
-
-        match leaf {
-            smt::smt_leaf::Leaf::Empty(leaf_index) => {
-                Ok(Self::new_empty(LeafIndex::new_max_depth(leaf_index)))
-            },
-            smt::smt_leaf::Leaf::Single(entry) => {
-                let (key, value): (Digest, Word) = entry.try_into()?;
-
-                Ok(SmtLeaf::new_single(key, value))
-            },
-            smt::smt_leaf::Leaf::Multiple(entries) => {
-                let domain_entries: Vec<(Digest, Word)> = try_convert(entries.entries)?;
-
-                Ok(SmtLeaf::new_multiple(domain_entries)?)
-            },
-        }
-    }
-}
-
-impl From<SmtLeaf> for smt::SmtLeaf {
-    fn from(smt_leaf: SmtLeaf) -> Self {
-        use smt::smt_leaf::Leaf;
-
-        let leaf = match smt_leaf {
-            SmtLeaf::Empty(leaf_index) => Leaf::Empty(leaf_index.value()),
-            SmtLeaf::Single(entry) => Leaf::Single(entry.into()),
-            SmtLeaf::Multiple(entries) => Leaf::Multiple(smt::SmtLeafEntries {
-                entries: convert(entries),
-            }),
-        };
-
-        Self { leaf: Some(leaf) }
-    }
-}
-
-impl TryFrom<smt::SmtLeafEntry> for (Digest, Word) {
-    type Error = ParseError;
-
-    fn try_from(entry: smt::SmtLeafEntry) -> Result<Self, Self::Error> {
-        let key: Digest =
-            entry.key.ok_or(smt::SmtLeafEntry::missing_field(stringify!(key)))?.try_into()?;
-        let value: Word = entry
-            .value
-            .ok_or(smt::SmtLeafEntry::missing_field(stringify!(value)))?
-            .try_into()?;
-
-        Ok((key, value))
-    }
-}
-
-impl From<(Digest, Word)> for smt::SmtLeafEntry {
-    fn from((key, value): (Digest, Word)) -> Self {
-        Self {
-            key: Some(key.into()),
-            value: Some(value.into()),
-        }
-    }
-}
-
-impl TryFrom<smt::SmtOpening> for SmtProof {
-    type Error = ParseError;
-
-    fn try_from(opening: smt::SmtOpening) -> Result<Self, Self::Error> {
-        let path: MerklePath = opening
-            .path
-            .ok_or(smt::SmtOpening::missing_field(stringify!(path)))?
-            .try_into()?;
-        let leaf: SmtLeaf = opening
-            .leaf
-            .ok_or(smt::SmtOpening::missing_field(stringify!(leaf)))?
-            .try_into()?;
-
-        Ok(SmtProof::new(path, leaf)?)
-    }
-}
-
-impl From<SmtProof> for smt::SmtOpening {
-    fn from(proof: SmtProof) -> Self {
-        let (path, leaf) = proof.into_parts();
-        Self {
-            path: Some(path.into()),
-            leaf: Some(leaf.into()),
         }
     }
 }
